+++ src/commands/git/pathHasGit/pathHasGit.ts
@@ -1,7 +1,0 @@
-import Path from 'path';
-import fse from 'fs-extra';
-
-/** True if has git, false if don't. */
-export function pathHasGit(path: string): Promise<boolean> {
-  return fse.pathExists(Path.join(path, '.git'));
-}