+++ src/commands/git/utils.ts
@@ -1,3 +1,11 @@
+import Path from 'path';
+import fse from 'fs-extra';
+
+/** True if has git, false if don't. */
+export function pathHasGit(path: string): Promise<boolean> {
+  return fse.pathExists(Path.join(path, '.git'));
+}
+
 /** Includes .git on end. */
 export function getRepositoryGitUrl(options: {
   owner: string;