+++ src/main/consts.ts
@@ -1,4 +1,0 @@
-/** Used in package.json.
- *
- * Is it require to have this before everything? */
-export const extensionIdentifier = 'githubRepoMgr';