+++ src/store/organization.ts
@@ -1,8 +1,9 @@
+import type { Octokit } from '@octokit/rest';
+import type { DirWithGitUrl } from '../commands/git/searchClonedRepos';
 import { getOrgRepos } from '../commands/github/getOrgRepos';
 import { getUserRepos } from '../commands/github/getUserRepos';
-import type { DirWithGitUrl } from '../commands/searchClonedRepos/searchClonedRepos';
 import type { Repository } from './repository';
-import { isRepoOnDisk } from './repository';
+import { isRepoNotOnDisk, isRepoOnDisk } from './repository';
 
 export enum OrgStatus {
   notLoaded,
@@ -21,14 +22,12 @@ type OrganizationConstructor = {
 };
 
 export class Organization {
-  name: string; // Used in notClonedRepo tree.
-  login: string;
-  status: OrgStatus;
-  repositories: Repository<boolean, 'user-is-member'>[] = [];
-  clonedRepos: Repository<true, 'user-is-member'>[] = [];
-  notClonedRepos: Repository<false, 'user-is-member'>[] = [];
+  name: string; // Pretty name
+  login: string; // Raw name
   /** If it is the user organization, = the repos belong to him. */
   isUserOrg: boolean;
+  clonedRepos: Repository<true, 'user-is-member'>[] = [];
+  notClonedRepos: Repository<false, 'user-is-member'>[] = [];
   /** If the user can create new repositories in this organization */
   userCanCreateRepositories: boolean;
 
@@ -37,38 +36,32 @@ export class Organization {
     this.name = args.name;
     this.isUserOrg = args.isUserOrg;
     this.userCanCreateRepositories = args.userCanCreateRepositories;
-    this.status = OrgStatus.notLoaded;
   }
 
   /** Call it after constructing it. */
-  async loadOrgRepos({ localRepos }: { localRepos: DirWithGitUrl[] }): Promise<void> {
-    try {
-      this.status = OrgStatus.loading;
+  async loadOrgRepos({
+    localRepos,
+    octokit,
+  }: {
+    localRepos: DirWithGitUrl[];
+    octokit: Octokit;
+  }): Promise<void> {
+    const repositories = this.isUserOrg
+      ? await getUserRepos({ octokit })
+      : await getOrgRepos({ login: this.login, octokit });
 
-      this.repositories = this.isUserOrg
-        ? await getUserRepos()
-        : await getOrgRepos(this.login);
-      // TODO splice localRepos so other orgs won't loop over it? Good for users/orgs with hundreds/thousands of repos.
-      localRepos.forEach((localRepo) => {
-        const repoMatch = this.repositories.find(
-          (repo) => repo.url === localRepo.gitUrl,
-        ) as Repository<true, 'user-is-member'> | undefined;
-        if (repoMatch) {
-          repoMatch.localPath = localRepo.dirPath;
-          repoMatch.dirty = 'unknown';
-        }
-      });
-      this.clonedRepos = this.repositories.filter((repo) =>
-        isRepoOnDisk(repo),
-      ) as Repository<true, 'user-is-member'>[];
-      this.notClonedRepos = this.repositories.filter(
-        (repo) => !isRepoOnDisk(repo),
-      ) as Repository<false, 'user-is-member'>[];
+    // Discover repos that are on disk
+    localRepos.forEach((localRepo) => {
+      const repositoryOnDisk = repositories.find(
+        (repo) => repo.url === localRepo.gitUrl,
+      ) as Repository<true, 'user-is-member'> | undefined;
+      if (repositoryOnDisk) {
+        repositoryOnDisk.localPath = localRepo.dirPath;
+        repositoryOnDisk.dirty = 'unknown';
+      }
+    });
 
-      this.status = OrgStatus.loaded;
-    } catch (err: any) {
-      this.status = OrgStatus.errorLoading;
-      throw new Error(err);
-    }
+    this.clonedRepos = repositories.filter(isRepoOnDisk);
+    this.notClonedRepos = repositories.filter(isRepoNotOnDisk);
   }
 }