+++ src/treeView/repositories/notClonedRepos.ts
@@ -1,7 +1,6 @@
 import vscode, { commands } from 'vscode';
-import { uiCloneTo } from '../../commandsUi/uiCloneTo';
 import type { Organization } from '../../store/organization';
-import { OrgStatus } from '../../store/organization';
+import { uiCloneTo } from '../../uiCommands/uiCloneTo';
 import { TreeItem } from '../treeViewBase';
 import { RepoItem } from './repoItem';
 
@@ -18,19 +17,6 @@ export function activateNotClonedRepos(): void {
   );
 }
 
-// Much like unused right now. Orgs will always be loaded here.
-function getEmptyOrgLabel(status: OrgStatus): string {
-  switch (status) {
-    case OrgStatus.errorLoading:
-      return 'Error loading';
-    case OrgStatus.notLoaded: // Same as loading.
-    case OrgStatus.loading:
-      return 'Loading...';
-    case OrgStatus.loaded:
-      return 'Empty';
-  }
-}
-
 export function getNotClonedTreeItem({
   userOrgs,
 }: {
@@ -39,20 +25,18 @@ export function getNotClonedTreeItem({
   const orgs: TreeItem[] = userOrgs.map((org) => {
     return new TreeItem({
       label: `${org.name}`,
-      children: org.repositories.length
-        ? org.notClonedRepos.map(
-            (repo) =>
-              new RepoItem({
-                repo,
-                contextValue: 'githubRepoMgr.context.notClonedRepo',
-                command: {
-                  // We wrap the repo in {} because we may call the cloneTo from the right click, and it passes the RepoItem.
-                  command: 'githubRepoMgr.commands.notClonedRepos.cloneTo',
-                  arguments: [{ repo }],
-                },
-              }),
-          )
-        : new TreeItem({ label: getEmptyOrgLabel(org.status) }),
+      children: org.notClonedRepos.map(
+        (repo) =>
+          new RepoItem({
+            repo,
+            contextValue: 'githubRepoMgr.context.notClonedRepo',
+            command: {
+              // We wrap the repo in {} because we may call the cloneTo from the right click, and it passes the RepoItem.
+              command: 'githubRepoMgr.commands.notClonedRepos.cloneTo',
+              arguments: [{ repo }],
+            },
+          }),
+      ),
       collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
     });
   });